server and serialport moved
